Refactor getDtsSnapshot and dotenv helpers
Resolves #37
Resolves #45
Resolves #46
Resolves #53
