+++ .github/workflows/ci.yml
@@ -10,8 +10,11 @@ jobs:
       - uses: actions/setup-node@v1
         with:
           node-version: 10.x
-      - name: Yarn install, build, and test
-        run: |
-          yarn install
-          yarn build
-          yarn test
+      - name: Install
+        run: yarn install --frozen-lockfile
+      - name: Lint
+        run: yarn lint
+      - name: Test
+        run: yarn test
+      - name: Build
+        run: yarn build
