+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "typescript-plugin-css-modules",
-  "version": "1.3.0",
+  "version": "1.4.0",
   "main": "lib/index.js",
   "author": "Brody McKee <mrmckeb@hotmail.com>",
   "license": "MIT",
@@ -25,12 +25,13 @@
   ],
   "scripts": {
     "build": "rm -rf ./lib && tsc",
+    "lint": "eslint './src/**/*.ts' --max-warnings 0 && yarn prettier './**/*.{json,md,ts,yml}' -c",
     "prepublishOnly": "yarn build",
     "test": "jest ./src"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "pretty-quick --staged",
+      "pre-commit": "lint-staged",
       "pre-push": "yarn test"
     }
   },
@@ -39,7 +40,22 @@
       "src/**/*.{ts}"
     ],
     "preset": "ts-jest",
-    "testEnvironment": "node"
+    "testEnvironment": "node",
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/fixtures/"
+    ]
+  },
+  "lint-staged": {
+    "./src/**/*.ts": [
+      "eslint --fix",
+      "prettier --write",
+      "git add"
+    ],
+    "./**/*.{json,md,yml}": [
+      "prettier --write",
+      "git add"
+    ]
   },
   "prettier": {
     "arrowParens": "always",
@@ -47,29 +63,38 @@
     "trailingComma": "all"
   },
   "dependencies": {
+    "dotenv": "^8.2.0",
     "icss-utils": "^4.1.1",
     "less": "^3.9.0",
     "lodash": "^4.17.14",
     "postcss": "^7.0.17",
+    "postcss-filter-plugins": "^3.0.1",
     "postcss-icss-selectors": "^2.0.3",
     "postcss-load-config": "^2.1.0",
     "reserved-words": "^0.1.2",
     "sass": "^1.22.4"
   },
   "devDependencies": {
+    "@types/icss-utils": "^4.1.0",
     "@types/jest": "^24.0.15",
     "@types/less": "^3.0.0",
     "@types/lodash": "^4.14.136",
-    "@types/node": "^10.0.0",
+    "@types/node": "^12.11.1",
+    "@types/postcss-load-config": "^2.0.0",
+    "@types/postcss-nested": "^4.1.0",
+    "@types/reserved-words": "^0.1.0",
     "@types/sass": "^1.16.0",
-    "husky": "^2.7.0",
+    "@typescript-eslint/eslint-plugin": "^2.4.0",
+    "@typescript-eslint/parser": "^2.4.0",
+    "eslint": "^6.5.1",
+    "eslint-config-prettier": "^6.4.0",
+    "husky": "^3.0.9",
     "jest": "^24.8.0",
+    "lint-staged": "^9.4.2",
     "postcss-import-sync2": "^1.1.0",
     "prettier": "^1.18.2",
-    "pretty-quick": "^1.11.1",
     "ts-jest": "^24.0.2",
-    "tslint": "^5.18.0",
-    "typescript": "^3.5.3"
+    "typescript": "^3.6.4"
   },
   "peerDependencies": {
     "typescript": "^3.0.0"
