--- src/@types/icss-utils.d.ts
@@ -1,12 +1,0 @@
-declare module 'icss-utils' {
-  import { Root } from 'postcss';
-  interface IICSSExports {
-    [exportName: string]: string;
-  }
-  export const extractICSS: (
-    css: Root,
-    removeRules?: boolean,
-  ) => {
-    icssExports: IICSSExports;
-  };
-}
