--- src/@types/postcss-nested.d.ts
@@ -1,5 +1,0 @@
-declare module 'postcss-nested' {
-  import { Plugin } from 'postcss';
-  const plugin: Plugin<any>;
-  export = plugin;
-}
