--- src/@types/reserved-words.d.ts
@@ -1,3 +1,0 @@
-declare module 'reserved-words' {
-  export const check: (string: string, esVersion?: string) => boolean;
-}
