--- src/@types/strip-css-singleline-comments.d.ts
@@ -1,4 +1,0 @@
-declare module 'strip-css-singleline-comments/sync' {
-  const strip: (s: string) => string;
-  export = strip;
-}
