--- src/config.ts
@@ -1,1 +1,0 @@
-export const packageName = 'typescript-plugin-css-modules';
