+++ src/helpers/__tests__/createMatchers.test.ts
@@ -2,11 +2,15 @@ import { createMatchers } from '../createMatchers';
 import { Options } from '../../options';
 import { Logger } from '../logger';
 
+const mockLogger: Logger = {
+  log: jest.fn(),
+  error: jest.fn(),
+};
+
 describe('utils / createMatchers', () => {
-  const logger: Logger = { log: jest.fn(), error: jest.fn() };
   it('should match `customMatcher` regexp', () => {
     const options: Options = { customMatcher: '\\.css$' };
-    const { isCSS, isRelativeCSS } = createMatchers(logger, options);
+    const { isCSS, isRelativeCSS } = createMatchers(mockLogger, options);
 
     expect(isCSS('./myfile.css')).toBe(true);
     expect(isCSS('./myfile.m.css')).toBe(true);
@@ -18,7 +22,7 @@ describe('utils / createMatchers', () => {
 
   it('should handle bad `customMatcher` regexp', () => {
     const options: Options = { customMatcher: '$([a' };
-    const { isCSS, isRelativeCSS } = createMatchers(logger, options);
+    const { isCSS, isRelativeCSS } = createMatchers(mockLogger, options);
 
     expect(isCSS('./myfile.module.css')).toBe(true);
     expect(isRelativeCSS('../folders/myfile.module.scss')).toBe(true);
