+++ src/helpers/__tests__/fixtures/customRenderer.js
@@ -1,0 +1,4 @@
+module.exports = (css, { fileName, logger }) => {
+  logger.log('Example log');
+  return `.${css} {}; .${fileName} {};`;
+};
