+++ src/options.ts
@@ -1,12 +1,24 @@
 import { Options as SassOptions } from 'sass';
+import { DotenvConfigOptions } from 'dotenv/types';
+import { Logger } from './helpers/logger';
+
+export interface PostCssOptions {
+  excludePlugins?: string[];
+  useConfig?: boolean;
+}
+
+export interface RendererOptions {
+  less?: Partial<Less.Options>;
+  sass?: Partial<SassOptions>;
+}
 
 export interface Options {
   camelCase?: CamelCaseOptions;
   customMatcher?: string;
-  renderOptions?: {
-    less?: Partial<Less.Options>;
-    scss?: Partial<SassOptions>;
-  };
+  customRenderer?: string;
+  dotEnvOptions?: DotenvConfigOptions;
+  postCssOptions?: PostCssOptions;
+  rendererOptions?: RendererOptions;
 }
 
 export type CamelCaseOptions =
@@ -15,3 +27,13 @@ export type CamelCaseOptions =
   | 'dashesOnly'
   | 'only'
   | undefined;
+
+export interface CustomRendererOptions {
+  fileName: string;
+  logger: Logger;
+}
+
+export type CustomRenderer = (
+  css: string,
+  Options: CustomRendererOptions,
+) => string;
