--- tslint.json
@@ -1,16 +1,0 @@
-{
-  "extends": "tslint:recommended",
-  "rules": {
-    "object-literal-key-quotes": [true, "as-needed"],
-    "quotemark": [true, "single"],
-    "trailing-comma": [
-      true,
-      {
-        "multiline": "always",
-        "singleline": "never",
-        "esSpecCompliant": true
-      }
-    ],
-    "variable-name": [true, "allow-leading-underscore"]
-  }
-}
